djangoprice/requirements.txt
black==25.1.0
celery~=5.5.3
Django~=4.2.23
django-celery-beat==2.8.1
django-environ~=0.12.0
djangorestframework~=3.16.0
psycopg2-binary~=2.9.10
redis==6.2.0
pytest~=8.4.1